Extract SubView (ImagesGridView) from MainView, which shows pictures of selected category (to reuse it in SearchScreenView).
