+++ Wallie/ResultScreenView/ResultScreenView.swift
@@ -1,0 +1,20 @@
+//
+//  ResultScreenView.swift
+//  Wallie
+//
+//  Created by Le Chris on 01.08.2023.
+//
+
+import SwiftUI
+
+struct ResultScreenView: View {
+    var body: some View {
+        Text("Hello, World!")
+    }
+}
+
+struct ResultScreenView_Previews: PreviewProvider {
+    static var previews: some View {
+        ResultScreenView()
+    }
+}
